Migrate fetch utility functions to the right place.
